+++ CHANGELOG.rst
@@ -19,6 +19,10 @@ Backward-incompatible changes:
 Changes:
 ^^^^^^^^
 
+- Add *repr_native_str* to ``structlog.processors.KeyValueRenderer`` and ``structlog.dev.ConsoleRenderer``.
+  This allows for human-readable non-ASCII output on Python 2 (``repr()`` on Python 2 haves like ``ascii()`` on Python 3 in that regard).
+  As per compatibility policy, it's on (original behavior) in ``KeyValueRenderer`` and off (humand-friendly behavior) in ``ConsoleRenderer``.
+  `#94 <https://github.com/hynek/structlog/issues/94>`_
 - Add *colors* argument to ``structlog.dev.ConsoleRenderer`` and make it the default renderer.
   `#78 <https://github.com/hynek/structlog/pull/78>`_